addressed comments and abstracted copying link logic into primary cta
